+++ glade_forms_devtools_extension/lib/src/widgets/model_card.dart
@@ -87,8 +87,8 @@ class ModelCard extends StatelessWidget {
                       label: '${model.childModels.length} models',
                       color: theme.colorScheme.primary,
                       isSmall: true,
-                    ),
-                  if (!model.isComposed || model.inputs.isNotEmpty)
+                    )
+                  else
                     _StatusChip(
                       label: '${model.inputs.length} inputs',
                       color: Colors.grey,